% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datapipeline_operations.R
\name{deactivate_pipeline}
\alias{deactivate_pipeline}
\title{Deactivates the specified running pipeline}
\usage{
deactivate_pipeline(pipelineId, cancelActive = NULL)
}
\arguments{
\item{pipelineId}{The ID of the pipeline.}

\item{cancelActive}{Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to \code{CANCELED}. If this value is false, the pipeline is deactivated after all running objects finish.}
}
\description{
Deactivates the specified running pipeline. The pipeline is set to the \code{DEACTIVATING} state until the deactivation process completes.
}
\details{
To resume a deactivated pipeline, use ActivatePipeline. By default, the pipeline resumes from the last completed execution. Optionally, you can specify the date and time to resume the pipeline.
}
\examples{

}
