% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datapipeline_operations.R
\name{delete_pipeline}
\alias{delete_pipeline}
\title{Deletes a pipeline, its pipeline definition, and its run history}
\usage{
delete_pipeline(pipelineId)
}
\arguments{
\item{pipelineId}{The ID of the pipeline.}
}
\description{
Deletes a pipeline, its pipeline definition, and its run history. AWS Data Pipeline attempts to cancel instances associated with the pipeline that are currently being processed by task runners.
}
\details{
Deleting a pipeline cannot be undone. You cannot query or restore a deleted pipeline. To temporarily pause a pipeline instead of deleting it, call SetStatus with the status set to \code{PAUSE} on individual components. Components that are paused by SetStatus can be resumed.
}
\examples{

}
