% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datapipeline_package.R
\docType{package}
\name{paws.datapipeline-package}
\alias{paws.datapipeline}
\alias{paws.datapipeline-package}
\title{paws.datapipeline: AWS Data Pipeline}
\description{
AWS Data Pipeline configures and manages a
data-driven workflow called a pipeline. AWS Data Pipeline handles the
details of scheduling and ensuring that data dependencies are met so
that your application can focus on processing the data.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
