% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datapipeline_operations.R
\name{add_tags}
\alias{add_tags}
\title{Adds or modifies tags for the specified pipeline}
\usage{
add_tags(pipelineId, tags)
}
\arguments{
\item{pipelineId}{The ID of the pipeline.}

\item{tags}{The tags to add, as key/value pairs.}
}
\description{
Adds or modifies tags for the specified pipeline.
}
\examples{

}
