% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datapipeline_operations.R
\name{evaluate_expression}
\alias{evaluate_expression}
\title{Task runners call \code{EvaluateExpression} to evaluate a string in the context of the specified object}
\usage{
evaluate_expression(
  pipelineId = "string",
  objectId = "string",
  expression = "string"
)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{objectId}{[required] The ID of the object.}

\item{expression}{[required] The expression to evaluate.}
}
\description{
Task runners call \code{EvaluateExpression} to evaluate a string in the context of the specified object. For example, a task runner can evaluate SQL queries stored in Amazon S3.
}
\examples{

}
